ai_resume_analyzer/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
SQLAlchemy>=2.0.0
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pdfminer.six>=20220524
python-docx>=1.0.0
Werkzeug>=3.0.0
